requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.8.0
eventlet==0.33.3
torch>=2.0.0
torchvision>=0.15.0
numpy==1.26.4
opencv-python>=4.5.0
pillow>=8.0.0
librosa==0.10.2.post1
tqdm
peft==0.15.1
transformers==4.52.3
scipy==1.14.0
xfuser==0.4.1
ftfy
einops
ffmpeg-python>=0.2.0 